1. If you spend too much time thinking about a thing, you'll never get it done

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,4 @@
-#include <stddef.h>
+#include <stdio.h>
 #include "function_pointers.h"
 
 /**
@@ -11,7 +11,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t num;
+	unsigned long int num;
 	int hold;
 
 	if (action == NULL)
